Updating dir names.
