--- student-blogs/Kick Off-2479/layout.css
@@ -1,84 +1,0 @@
-html,
-body {
-    margin: 5px;
-    height: 100%;
-    box-sizing: border-box;
-
-}
-
-* {
-    box-sizing: inherit;
-}
-
-main {
-    display: flex;
-    flex-direction: column;
-}
-
-body {
-    font-family: "Century Gothic", sans-serif;
-    line-height: 2;
-    color: #333;
-}
-
-h1 {
-    text-align: center;
-    font-size: 30px;
-    color: brown;
-    text-shadow: 1px 1px 1px black;
-}
-
-p {
-    border: 3px solid gray;
-    border-radius: 20px;
-    border-color: brown;
-    text-shadow: 0.5px 1px 1px;    
-    color: #333;
-    padding: 1em;
-}
-
-strong {
-    color: brown;
-}
-
-div {
-    margin-top: 10px;
-    border: 1px solid Gainsboro;
-    padding: 10px;
-}
-
-img {
-    width: 100%;
-    height: 50%;
-    border-radius: 22px;
-}
-
-
-@media screen and (min-width: 992px) {
-    body {
-        width: 60%;
-        margin: auto;
-        line-height: 1.5;
-    }
-
-
-    .containers {
-        display: grid;
-        grid-template-rows: auto;
-        justify-content: space-evenly;
-        align-items: center;
-        justify-items: center;
-        gap: 5px;
-    }
-    
-    img{
-        width: 100%;
-        height: 100%;
-    }
-
-    h1 {
-        font-size: 30px;
-        color: brown;
-        text-shadow: 1px 1px 1px black;
-    }
-}
