+++ student-blogs/Kick-Off-2479/index.html
@@ -1,0 +1,105 @@
+<!DOCTYPE html>
+<html lang="es">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>My curriculum vitae</title>
+    <link rel="stylesheet" href="layout.css">
+</head>
+
+<body>
+    <main>
+        <div class="containers">
+            <img id="picture_1"
+                src="https://images.unsplash.com/photo-1551753677-dff56803a44f?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=2070&q=80">
+            <h1>Who was Aristotle?</h1>
+            <p>Ancient Greek philosopher and scientist, one of the greatest intellectual figures of Western
+                history. He
+                was the author of a philosophical and scientific system that became the framework and
+                vehicle
+                for both
+                Christian Scholasticism and medieval Islamic philosophy. Even after the intellectual
+                revolutions
+                of the
+                Renaissance, the Reformation, and the Enlightenment, Aristotelian concepts remained embedded
+                in
+                Western
+                thinking.</p>
+            <img
+                src="https://images.unsplash.com/photo-1505664194779-8beaceb93744?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1470&q=80">
+            <h1>What is modern philosophy?</h1>
+            <p>Is philosophy developed in the modern era and associated with modernity. It is not a specific
+                doctrine or
+                school (and thus should not be confused with Modernism), although there are certain
+                assumptions
+                common
+                to much of it, which helps to distinguish it from earlier philosophy.
+
+                The 17th and early 20th centuries roughly mark the beginning and the end of modern
+                philosophy.
+                How much
+                of the Renaissance should be included is a matter for dispute; likewise modernity may or may
+                not
+                have
+                ended in the twentieth century and been replaced by postmodernity. How one decides these
+                questions will
+                determine the scope of one's use of the term "modern philosophy."</p>
+            <img
+                src="https://images.unsplash.com/photo-1620662831351-9f68f76d0b9a?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1170&q=80">
+            <h1>Should we think? / Overthink</h1>
+            <p>There are no “shoulds” when it comes to mindset.
+
+                I believe the better path is to practice getting out of your head and into your body in the
+                now.
+
+                Most people miss out on a lot of living because they indulge and overthink their thoughts.
+
+                I challenge you to try to take a step back from your thoughts and just observe them. Are
+                they
+                positive
+                or negative? Do they make you feel good, bad, bored? Do you obsess on the past or does the
+                past
+                only
+                come up occasionally? How often are your thoughts oriented to your future? Are your future
+                thoughts
+                about success, relationships, your interests? Is the tone of your future positive or
+                negative in
+                your
+                thoughts? Or does your future start out rosy only to be hijacked by doom and gloom when you
+                stop
+                paying
+                attention? What recurring thoughts do you notice? Are the thoughts you observe serving you
+                positively or
+                are they suffocating and oppressive?
+
+                Above all if you want to change the quality of your thoughts pay attention to your self
+                talk.
+                Are you
+                patient, supportive, or your own best cheer leader? Are you short with yourself, do your
+                thoughts pounce
+                on your missteps, or does your inner dialogue belittle and begrudge the things you do?</p>
+            <img
+                src="https://images.unsplash.com/photo-1564648351416-3eec9f3e85de?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1074&q=80">
+            <h1>What is perspective?</h1>
+            <p>What is the definition of perspective? In general, perspective can refer to a person's
+                outlook or
+                view on
+                something, such as an event, situation, or person. So, what is perspective in literature? By
+                definition,
+                perspective in literature is the narrator's outlook or view on events, characters, and the
+                world. Just
+                as people's perspectives are influenced by a particular set of circumstances in real life,
+                narrators'
+                perspectives are influenced by who they are, what they know, what they have experienced, and
+                so
+                on.
+                Therefore, a narrator's perspective determines how a story is told to the reader.</p>
+        </div>
+        </div>
+        </section>
+        </article>
+    </main>
+</body>
+
+</html>
