--- student-cvs/Kick Off-2479/main.mjs
@@ -1,7 +1,0 @@
-function handleSubmit(event) {
-    event.preventDefault();
-    console.log("Hola mundo");
-}
-    
-const $form =document.querySelector("form");
-$form.addEventListener("submit", handleSubmit);
