--- student-cvs/Kick Off-2479/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "kick-off-2479",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {    
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "vite": "^4.2.1"
-  }
-}
