--- student-forms/Kick Off-2479/bootstrap-setup/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "bootstrap-setup",
-  "version": "1.0.0",
-  "main": "index.js",
-  "scripts": {
-    "start": "vite",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@types/bootstrap": "^5.2.6",
-    "sass": "^1.60.0",
-    "vite": "^4.2.1"
-  },
-  "dependencies": {
-    "@popperjs/core": "^2.11.7",
-    "bootstrap": "^5.2.3"
-  },
-  "description": ""
-}
