--- student-forms/Kick Off-2479/bootstrap-setup/src/scss/styles.scss
@@ -1,2 +1,0 @@
-// Import all of Bootstrap's CSS
-@import "bootstrap/scss/bootstrap";
