--- student-forms/Kick Off-2479/bootstrap-setup/vite.config.js
@@ -1,12 +1,0 @@
-const path = require('path')
-
-export default {
-  root: path.resolve(__dirname, 'src'),
-  build: {
-    outDir: '../dist'
-  },
-  server: {
-    port: 8080,
-    hot: true
-  }
-}
