+++ student-forms/Kick-Off-2479/bootstrap-setup/src/js/main.js
@@ -1,0 +1,10 @@
+// Import our custom CSS
+import '../scss/styles.scss'
+
+// Import all of Bootstrap's JS
+import * as bootstrap from 'bootstrap'
+
+import Alert from 'bootstrap/js/dist/alert';
+
+// or, specify which plugins you need:
+import { Tooltip, Toast, Popover } from 'bootstrap';
